fix(statistics): change structure of statistics components, limit filters based on type
