+++ src/features/courses/chartConstant.ts
@@ -1,0 +1,19 @@
+export const kChartAnimationDuration = 200; // ms
+
+export const emptyChartData = [
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+  { value: 0, frontColor: 'transparent' },
+];
